+++ backend/Tasque.Core.Common/DTO/PartialModels/CreateOrganization.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Tasque.Core.Common.IncomeModels
+namespace Tasque.Core.Common.PartialModels
 {
     public class CreateOrganization
     {
